chore: more modular python code
